[refactoring] According to CorpusProcessAnnotationLogs package move
